Add sponsors
Sponsors are interspersed between each content section and are given
an amount of space proportional to how much they've sponsored us for.

Specifically (as measured in Bootstrap columns):

* IPO - 6
* Mezzanine - 4
* Series A - 3
* Seed - 2

I tried a few other configurations, giving Seed 1 for example or IPO 12
for example, but they were found to be unworkable either because they
were too big and seemed intrusive or too small so as to be unreadable.