Restructure index
The masthead was rendering incorrectly in the DOM since it was inside
any block. I solved this by making a header and footer block before and
after the container respectively which can be overriden to keep all
content within body. While I was at it, I also moved out `footer.html`
into the footer block.