Add navbar
Adds a navbar to the site. The navbar sticks to the top once its
scrolled by and tracks where you are at on the page. As of now, it's
not very useful as our sections are very small, but so it make make
sense to remove this later, but I think it would be a good goal to do
what it takes to keep this around.