Update the headline and add a pitch
The new headline is, "Eat. Sleep. Code." and the pitch is:

"PennApps is a 48-hour application development competition at the
University of Pennsylvania where the top student engineers, designers,
and entrepreneurs come together to showcase technical achievements and
compete for cash prizes."

The motivation behind the headline is to describe PennApps as briefly as
possible. I can't think of anything more terse than the three words that
make up our tagline. I've decided to remove the strike-through for two
reasons: 1. It looks horrible. 2. It doesn't really matter much anyhow,
as people do sleep, and the strike-through takes more mental effort to
process.

The motivation behind the pitch is to describe the who, what, when,
where, and why of PennApps as briefly as possible. I think I covered it,
but I'm open for revisions.

Together the headline and pitch let visitors know exactly what PennApps
is within 2-3 seconds, and provides them with all the necessary
information to decide whether or not they want to sign up.